+++ CHANGELOG.md
@@ -29,6 +29,7 @@ height. Use `-d` to specify the delta number of blocks back to compare to.
 - Relic and thus blspy would crash on processors older than Haswell as they don't support lzc.
 - chia-check-plots works with plot root (the same as harvester)
 - Some networking errors are no longer printed
+- Blocks with compact proof of time are now updated into database.
 
 ## [1.0beta5] aka Beta 1.5 - 2020-05-05
 ### Added